feat: extract core logic
