+++ src/builtin-addons/core/template-context-provider.ts
@@ -2,9 +2,9 @@ import * as path from 'path';
 import * as fs from 'fs';
 import { CompletionItem, CompletionItemKind } from 'vscode-languageserver';
 
-import { isModuleUnificationApp, podModulePrefixForRoot, pureComponentName } from '../utils/layout-helpers';
+import { isModuleUnificationApp, podModulePrefixForRoot, pureComponentName } from '../../utils/layout-helpers';
 
-import { log } from '../utils/logger';
+import { log } from '../../utils/logger';
 
 import { extractComponentInformationFromMeta, processJSFile, processTemplate } from 'ember-meta-explorer';
